requirements.txt
# 本地AI助手 - Python依赖

# ===== 核心框架 =====
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# ===== 数据库 =====
sqlalchemy>=2.0.0

# ===== 配置管理 =====
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ===== HTTP客户端 =====
httpx>=0.24.0
requests>=2.28.0

# ===== 表单处理 =====
python-multipart>=0.0.6

# ===== 文档解析 =====
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
chardet>=5.0.0
Pillow>=10.0.0
pdf2image>=1.16.0

# ===== OCR文字识别 =====
rapidocr-onnxruntime>=1.3.0
